LPS-112247 Move logic to each of their implementations
